+++ index.php
@@ -31,6 +31,8 @@ require_once('includes/conexion.inc.php');
             }
         }
 
+        //Comprobaciones de los datos
+
         if (!preg_match($exprNombre, $_POST["nombre"]) && !isset($errores["nombre"])) {
             $errores["nombre"] = '<p class="error">El nombre solo recibe de 3 a 15 letras y números.</p><br>';
         }
@@ -49,9 +51,11 @@ require_once('includes/conexion.inc.php');
 
 
 
+        //Si no hay errores se procede a realizar la inserción o actualización de los datos
         if (!$errores) {
             $conexion = conectar();
-
+            
+            //En caso de que se haya enviado el codigo del grupo se realizará un update, en caso contrario un insert
             if (!is_null($conexion)) {
                 if (!isset($_POST['codigo'])) {
 
@@ -84,13 +88,13 @@ require_once('includes/conexion.inc.php');
     }
 
     //En caso de que se esté borrando un dato entrará aquí
-    if (isset($_GET['accion'])) {
+    if (isset($_GET['accion']) && $_GET['accion'] == 'borrar') {
         $conexion = conectar();
 
         if (!is_null($conexion)) {
             $consulta = $conexion->prepare('DELETE FROM grupos WHERE codigo=?');
 
-            $consulta->bindParam(1, $_POST["codigo"]);
+            $consulta->bindParam(1, $_GET["codigo"]);
         }
         try {
             $consulta->execute();
@@ -99,6 +103,7 @@ require_once('includes/conexion.inc.php');
 
         unset($conexion);
         unset($consulta);
+
         header('Location: index.php');
     }
 
@@ -108,10 +113,19 @@ require_once('includes/conexion.inc.php');
 <body>
     <?php
     include_once('includes/cabecera.inc.php');
+
+    //En caso de que se quiera borrar un grupo entrará en esta confirmación
+    if (isset($_GET['accion']) && $_GET['accion'] == 'confirmar') {
+        echo '<div class="borrar">';
+        echo '  <h2>¿Estás seguro de que quieres borrar este grupo?</h2>';
+        echo '  <a href="index.php?codigo=' . $_GET["codigo"] . '&accion=borrar"">Borrar</a>';
+        echo '</div>';
+    }
     ?>
     <div class="grupos">
         <ol>
             <?php
+            //Generación de la lista de grupos
             $conexion = conectar();
 
             if (!is_null($conexion)) {
@@ -121,7 +135,7 @@ require_once('includes/conexion.inc.php');
                     echo '<li>
                             <a href="grupo.php?grupo=' . $grupo["codigo"] . '">' . " " . $grupo["nombre"] . '</a>
                             <a href="index.php?codigo=' . $grupo["codigo"] . '"><img src="img/editar.png" alt="' . $grupo["nombre"] . '_icono_editar"></a>
-                            <a href="index.php?codigo=' . $grupo["codigo"] . '&accion=borrar""><img src="img/borrar.png" alt="' . $grupo["nombre"] . '_icono_borrar"></a>
+                            <a href="index.php?codigo=' . $grupo["codigo"] . '&accion=confirmar""><img src="img/borrar.png" alt="' . $grupo["nombre"] . '_icono_borrar"></a>
                         </li>';
                 }
             }
@@ -132,6 +146,7 @@ require_once('includes/conexion.inc.php');
     </div>
     <div>
         <?php
+        //En caso de que se haya enviado el codigo del grupo se mostrará el formulario con los datos del grupo
         $conexion = conectar();
 
         if (!is_null($conexion) && isset($_GET['codigo'])) {
@@ -172,11 +187,11 @@ require_once('includes/conexion.inc.php');
             <label for="inicio">Año de inicio</label><br>
             <input name="inicio" id="inicio" value="<?= $_POST["inicio"] ?? "" ?>"><br>
 
-            <input type="hidden" name="codigo" value="<?= $_GET["codigo"] ?? "" ?>">
-
             <?php echo isset($errores["inicio"]) ? $errores["inicio"] : "" ?>
             <?php
+            //En caso de que se haya enviado el codigo del grupo se almacenará en un input hidden
             if (isset($_GET['codigo'])) {
+                echo '<input type="hidden" name="codigo" value="' . $_GET["codigo"] . '">';
                 echo '<input type="submit" value="Editar">';
                 echo '<a href="index.php">Cancelar</a>';
             } else {
